api/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
